Adding script to vet output of simulations.
This script shouldn't be necessary, but due to the scheduler on our
cluster misbehaving, some of ecoevolity analyses of simulated datasets
get killed. This script checks the output of all analyses to find those
that were killed and then reports the path to the job submission scripts
that need to be re-run.